Sync Fork from Upstream Repo

+++ pkg/jobs/registry.go
@@ -431,8 +431,8 @@ func (r *Registry) CreateJobWithTxn(ctx context.Context, record Record, txn *kv.
 	if !r.startUsingSQLLivenessAdoption(ctx) {
 		// TODO(spaskob): remove in 20.2 as this code path is only needed while
 		// migrating to 20.2 cluster.
-		if err := j.WithTxn(txn).deprecatedInsert(
-			ctx, r.makeJobID(), r.deprecatedNewLease(), s,
+		if err := j.deprecatedInsert(
+			ctx, txn, r.makeJobID(), r.deprecatedNewLease(), s,
 		); err != nil {
 			return nil, err
 		}
@@ -477,8 +477,8 @@ func (r *Registry) CreateAdoptableJobWithTxn(
 	// in the cluster) to adopt this job at a later time.
 	lease := &jobspb.Lease{NodeID: invalidNodeID}
 
-	if err := j.WithTxn(txn).deprecatedInsert(
-		ctx, r.makeJobID(), lease, nil,
+	if err := j.deprecatedInsert(
+		ctx, txn, r.makeJobID(), lease, nil,
 	); err != nil {
 		return nil, err
 	}
@@ -505,11 +505,6 @@ func (r *Registry) CreateStartableJobWithTxn(
 	if err != nil {
 		return nil, err
 	}
-	// The job itself must not hold on to this transaction. We ensure in Start()
-	// that the transaction used to create the job is committed. When jobs hold
-	// onto transactions they use the transaction in methods which modify the job.
-	// On the whole this pattern is bug-prone and hard to reason about.
-	j.WithTxn(nil)
 	resumer, err := r.createResumer(j, r.settings)
 	if err != nil {
 		return nil, err
@@ -517,11 +512,19 @@ func (r *Registry) CreateStartableJobWithTxn(
 	// Construct a context which contains a tracing span that follows from the
 	// span in the parent context. We don't directly use the parent span because
 	// we want independent lifetimes and cancellation. For the same reason, we
-	// don't use the Context returned by ForkCtxSpan.
+	// don't use the Context returned by ForkSpan.
 	resumerCtx, cancel := r.makeCtx()
-	_, span := tracing.ForkCtxSpan(ctx, "job")
+	_, span := tracing.ForkSpan(ctx, "job")
 	if span != nil {
 		resumerCtx = tracing.ContextWithSpan(resumerCtx, span)
+
+		// This trace span unfortunately is sometimes never finished.
+		// As a hack/workaround, finish it now so that it leaves the
+		// tracer registry.
+		//
+		// Remove this when this issue is fixed:
+		// https://github.com/cockroachdb/cockroach/issues/60671
+		span.Finish()
 	}
 
 	if r.startUsingSQLLivenessAdoption(ctx) {
@@ -564,7 +567,7 @@ func (r *Registry) LoadJobWithTxn(ctx context.Context, jobID int64, txn *kv.Txn)
 		id:       &jobID,
 		registry: r,
 	}
-	if err := j.WithTxn(txn).load(ctx); err != nil {
+	if err := j.load(ctx, txn); err != nil {
 		return nil, err
 	}
 	return j, nil
@@ -580,7 +583,7 @@ func (r *Registry) UpdateJobWithTxn(
 		id:       &jobID,
 		registry: r,
 	}
-	return j.WithTxn(txn).Update(ctx, updateFunc)
+	return j.Update(ctx, txn, updateFunc)
 }
 
 // DefaultCancelInterval is a reasonable interval at which to poll this node
@@ -898,30 +901,31 @@ func (r *Registry) cleanupOldJobs(ctx context.Context, olderThan time.Time) erro
 // previous page).
 func (r *Registry) cleanupOldJobsPage(
 	ctx context.Context, olderThan time.Time, minID int64, pageSize int,
-) (done bool, maxID int64, _ error) {
+) (done bool, maxID int64, retErr error) {
 	const stmt = "SELECT id, payload, status, created FROM system.jobs " +
 		"WHERE created < $1 AND id > $2 " +
 		"ORDER BY id " + // the ordering is important as we keep track of the maximum ID we've seen
 		"LIMIT $3"
-	rows, err := r.ex.Query(ctx, "gc-jobs", nil /* txn */, stmt, olderThan, minID, pageSize)
+	it, err := r.ex.QueryIterator(ctx, "gc-jobs", nil /* txn */, stmt, olderThan, minID, pageSize)
 	if err != nil {
 		return false, 0, err
 	}
-	log.VEventf(ctx, 2, "read potentially expired jobs: %d", len(rows))
-
-	if len(rows) == 0 {
-		return true, 0, nil
-	}
-	// Track the highest ID we encounter, so it can serve as the bottom of the
-	// next page.
-	maxID = int64(*(rows[len(rows)-1][0].(*tree.DInt)))
-	// If we got as many rows as we asked for, there might be more.
-	morePages := len(rows) == pageSize
-
+	// We have to make sure to close the iterator since we might return from the
+	// for loop early (before Next() returns false).
+	defer func() {
+		closeErr := it.Close()
+		if retErr == nil {
+			retErr = closeErr
+		}
+	}()
 	toDelete := tree.NewDArray(types.Int)
-	toDelete.Array = make(tree.Datums, 0, len(rows))
 	oldMicros := timeutil.ToUnixMicros(olderThan)
-	for _, row := range rows {
+
+	var ok bool
+	var numRows int
+	for ok, err = it.Next(ctx); ok; ok, err = it.Next(ctx) {
+		numRows++
+		row := it.Cur()
 		payload, err := UnmarshalPayload(row[1])
 		if err != nil {
 			return false, 0, err
@@ -941,6 +945,14 @@ func (r *Registry) cleanupOldJobsPage(
 			toDelete.Array = append(toDelete.Array, row[0])
 		}
 	}
+	if err != nil {
+		return false, 0, err
+	}
+	if numRows == 0 {
+		return true, 0, nil
+	}
+
+	log.VEventf(ctx, 2, "read potentially expired jobs: %d", numRows)
 	if len(toDelete.Array) > 0 {
 		log.Infof(ctx, "cleaning up expired job records: %d", len(toDelete.Array))
 		const stmt = `DELETE FROM system.jobs WHERE id = ANY($1)`
@@ -955,6 +967,12 @@ func (r *Registry) cleanupOldJobsPage(
 				len(toDelete.Array), nDeleted)
 		}
 	}
+	// If we got as many rows as we asked for, there might be more.
+	morePages := numRows == pageSize
+	// Track the highest ID we encounter, so it can serve as the bottom of the
+	// next page.
+	lastRow := it.Cur()
+	maxID = int64(*(lastRow[0].(*tree.DInt)))
 	return !morePages, maxID, nil
 }
 
@@ -989,12 +1007,12 @@ func (r *Registry) CancelRequested(ctx context.Context, txn *kv.Txn, id int64) e
 			// safest way for now (i.e., without a larger jobs/schema change refactor)
 			// is to hack this up with a string comparison.
 			if payload.Type() == jobspb.TypeSchemaChange && !strings.HasPrefix(payload.Description, "ROLL BACK") {
-				return job.WithTxn(txn).cancelRequested(ctx, nil)
+				return job.cancelRequested(ctx, txn, nil)
 			}
 		}
 		return err
 	}
-	return job.WithTxn(txn).cancelRequested(ctx, nil)
+	return job.cancelRequested(ctx, txn, nil)
 }
 
 // PauseRequested marks the job with id as paused-requested using the specified txn (may be nil).
@@ -1007,7 +1025,7 @@ func (r *Registry) PauseRequested(ctx context.Context, txn *kv.Txn, id int64) er
 	if pr, ok := resumer.(PauseRequester); ok {
 		onPauseRequested = pr.OnPauseRequest
 	}
-	return job.WithTxn(txn).pauseRequested(ctx, onPauseRequested)
+	return job.pauseRequested(ctx, txn, onPauseRequested)
 }
 
 // Succeeded marks the job with id as succeeded.
@@ -1016,7 +1034,7 @@ func (r *Registry) Succeeded(ctx context.Context, txn *kv.Txn, id int64) error {
 	if err != nil {
 		return err
 	}
-	return job.WithTxn(txn).succeeded(ctx, nil)
+	return job.succeeded(ctx, txn, nil)
 }
 
 // Failed marks the job with id as failed.
@@ -1025,7 +1043,7 @@ func (r *Registry) Failed(ctx context.Context, txn *kv.Txn, id int64, causingErr
 	if err != nil {
 		return err
 	}
-	return job.WithTxn(txn).failed(ctx, causingError, nil)
+	return job.failed(ctx, txn, causingError, nil)
 }
 
 // Unpause changes the paused job with id to running or reverting using the
@@ -1035,7 +1053,7 @@ func (r *Registry) Unpause(ctx context.Context, txn *kv.Txn, id int64) error {
 	if err != nil {
 		return err
 	}
-	return job.WithTxn(txn).unpaused(ctx)
+	return job.unpaused(ctx, txn)
 }
 
 // Resumer is a resumable job, and is associated with a Job object. Jobs can be
@@ -1142,7 +1160,7 @@ func (r *Registry) stepThroughStateMachine(
 		}
 		resumeCtx := logtags.AddTag(ctx, "job", *job.ID())
 		if payload.StartedMicros == 0 {
-			if err := job.started(ctx); err != nil {
+			if err := job.started(ctx, nil /* txn */); err != nil {
 				return err
 			}
 		}
@@ -1189,7 +1207,7 @@ func (r *Registry) stepThroughStateMachine(
 		return errors.NewAssertionErrorWithWrappedErrf(jobErr,
 			"job %d: unexpected status %s provided to state machine", *job.ID(), status)
 	case StatusCanceled:
-		if err := job.canceled(ctx, nil); err != nil {
+		if err := job.canceled(ctx, nil /* txn */, nil /* fn */); err != nil {
 			// If we can't transactionally mark the job as canceled then it will be
 			// restarted during the next adopt loop and reverting will be retried.
 			return errors.Wrapf(err, "job %d: could not mark as canceled: %v", *job.ID(), jobErr)
@@ -1200,7 +1218,7 @@ func (r *Registry) stepThroughStateMachine(
 			return errors.NewAssertionErrorWithWrappedErrf(jobErr,
 				"job %d: successful bu unexpected error provided", *job.ID())
 		}
-		if err := job.succeeded(ctx, nil); err != nil {
+		if err := job.succeeded(ctx, nil /* txn */, nil /* fn */); err != nil {
 			// If it didn't succeed, we consider the job as failed and need to go
 			// through reverting state first.
 			// TODO(spaskob): this is silly, we should remove the OnSuccess hooks and
@@ -1210,7 +1228,7 @@ func (r *Registry) stepThroughStateMachine(
 		}
 		return nil
 	case StatusReverting:
-		if err := job.reverted(ctx, jobErr, nil); err != nil {
+		if err := job.reverted(ctx, nil /* txn */, jobErr, nil /* fn */); err != nil {
 			// If we can't transactionally mark the job as reverting then it will be
 			// restarted during the next adopt loop and it will be retried.
 			return errors.Wrapf(err, "job %d: could not mark as reverting: %s", *job.ID(), jobErr)
@@ -1257,7 +1275,7 @@ func (r *Registry) stepThroughStateMachine(
 			return errors.NewAssertionErrorWithWrappedErrf(jobErr,
 				"job %d: has StatusFailed but no error was provided", *job.ID())
 		}
-		if err := job.failed(ctx, jobErr, nil); err != nil {
+		if err := job.failed(ctx, nil /* txn */, jobErr, nil /* fn */); err != nil {
 			// If we can't transactionally mark the job as failed then it will be
 			// restarted during the next adopt loop and reverting will be retried.
 			return errors.Wrapf(err, "job %d: could not mark as failed: %s", *job.ID(), jobErr)
